[extimg] move ExternalImagesStackPluginBase
