+++ PyMca5/PyMcaGui/pymca/ExternalImagesStackPluginBase.py
